Changes to FlashWriteSize
Looking at esptool.py it looks like all STUBS have 0x4000 as a flash write size. Updating this along with getFlashWriteSize to be able to flash faster. A bigger data block means less commands to be sent to the STUB.
Before this change it was flashing in blocks of 0x200.